+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Root command: don't swallow errors fetching dependencies. Retry in case of error.
+
 ## [0.26.3] - 2026-03-12
 
 ### Fixed